requirement.txt
PyJWT==2.8.0
bcrypt==4.1.2
python-dotenv